+++ src/utils/pdf/index.ts
@@ -1,3 +1,5 @@
 // PDF utilities exports
 export * from './imageToPDF';
 export * from './pdfMerger';
+export * from './pdfOrganizer';
+export * from './pdfRenderer';